refactor: use low level r2dbc databaseClient instead of repository
also add parent and children fetching methods

+++ graphql/src/main/kotlin/com/github/ben_lc/sapin/graphql/type/EntityExtensions.kt
@@ -7,7 +7,7 @@ fun LocationEntity.toGraphqlType(): Location =
     Location(
         id = this.id.toString(),
         name = this.name,
-        level = this.level,
+        level = this.level.toInt(),
         levelLocalName = this.levelLocalName,
         levelLocalNameEn = this.levelLocalNameEn,
         isoId = this.isoId)
